JToric – toric geometry in Julia

JToric – toric geometry in Julia

Goal

Toric Varieties

Toric Divisors

Goal

The goal of the package JToric is to make computations on toric geometry accessible in the OSCAR computer algebra system. As such, we follow the book Toric Varieties by David A. Cox, John B. Little and Henry K. Schenck]. Currently, this project is work-in-progress. Our goal is to ensure that one can perform all computations of Appendix B.

Generalities

JToric.version — Constant

version

The version number of the loaded JToric. Please mention this number in any bug report.

Toric Varieties

Constructors

JToric.NormalToricVariety — Type

NormalToricVariety( r::Matrix{Int}, c::Vector{Vector{Int}} )

Construct the normal toric variety whose fan has ray generators r and maximal cones c.

Examples

julia> NormalToricVariety( [-1 5; 0 1; 1 0; 0 -1], [[1,2],[2,3],[3,4],[4,1]] )
NormalToricVariety(GAP: <A toric variety of dimension 2>, Polymake.BigObjectAllocated(Ptr{Nothing} @0x0000561365a4f3d0))

NormalToricVariety( v::GapObj )

Construct the Julia wrapper for a GAP toric variety v.

JToric.projective_space — Function

projective_space( d::Int )

Construct the projective space of dimension d.

Examples

julia> projective_space( 2 )
NormalToricVariety(GAP: <A projective toric variety of dimension 2>, Polymake.BigObjectAllocated(Ptr{Nothing} @0x0000562fbcc21b70))

JToric.hirzebruch_surface — Function

hirzebruch_surface( r::Int )

Constructs the r-th Hirzebruch surface.

Examples

julia> hirzebruch_surface( 5 )
NormalToricVariety(GAP: <A toric variety of dimension 2>, Polymake.BigObjectAllocated(Ptr{Nothing} @0x0000562fbcc21b70))

JToric.del_pezzo — Function

delPezzo( b::Int )

Constructs the delPezzo surface with b blowups for b at most 3.

Examples

julia> del_pezzo( 3 )
NormalToricVariety(GAP: <A toric variety of dimension 2>, Polymake.BigObjectAllocated(Ptr{Nothing} @0x0000562fbcc21b70))

Conversion among GAP and Polymake toric varieties

JToric.ntv_gap2polymake — Function

ntv_gap2polymake( v::GapObj )

Convert a GAP toric variety v into a Polymake toric variety.

JToric.ntv_polymake2gap — Function

ntv_polymake2gap( v::Polymake.BigObject )

Convert a Polymake toric variety v into a GAP toric variety.

Properties of toric varieties

JToric.is_normal_variety — Function

is_normal_variety( v::NormalToricVariety )

Checks if the normal toric variety v is normal. (This function is somewhat tautological at this point.)

Examples

julia> is_normal_variety(projective_space( 2 ))
true

JToric.is_affine — Function

is_affine( v::NormalToricVariety )

Checks if the normal toric variety v is affine.

Examples

julia> is_normal_variety( projective_space( 2 ) )
false

JToric.is_projective — Function

is_projective( v::NormalToricVariety )

Checks if the normal toric variety v is projective, i.e. if the fan of v is the the normal fan of a polytope.

Examples

julia> is_projective( projective_space( 2 ) )
true

JToric.is_smooth — Function

is_smooth( v::NormalToricVariety )

Checks if the normal toric variety v is smooth.

Examples

julia> is_smooth( projective_space( 2 ) )
true

JToric.is_complete — Function

is_complete( v::NormalToricVariety )

Checks if the normal toric variety v is complete.

Examples

julia> is_complete( projective_space( 2 ) )
true

JToric.has_torusfactor — Function

has_torusfactor( v::NormalToricVariety )

Checks if the normal toric variety v has a torus factor.

Examples

julia> has_torusfactor( projective_space( 2 ) )
false

JToric.is_orbifold — Function

is_orbifold( v::NormalToricVariety )

Checks if the normal toric variety v is an orbifold.

Examples

julia> is_orbifold( projective_space( 2 ) )
true

JToric.is_simplicial — Function

is_simplicial( v::NormalToricVariety )

Checks if the normal toric variety v is simplicial. Hence, this function works just as is_orbifold. It is implemented for user convenience.

Examples

julia> is_simplicial( projective_space( 2 ) )
true

JToric.is_isomorphic_to_projective_space — Function

is_isomorphic_to_projective_space( v::NormalToricVariety )

Checks if the normal toric variety v is isomorphic to projective space.

Examples

julia> is_isomorphic_to_projective_space( projective_space( 2 ) )
true

JToric.is_direct_product_of_projective_spaces — Function

is_direct_product_of_projective_spaces( v::NormalToricVariety )

Checks if the normal toric variety v is isomorphic to a direct product of projective space.

Examples

julia> is_direct_product_of_projective_spaces( projective_space( 2 ) )
true

JToric.is_gorenstein — Function

is_gorenstein( v::NormalToricVariety )

Checks if the normal toric variety v is Gorenstein.

Examples

julia> is_gorenstein( projective_space( 2 ) )
true

JToric.is_q_gorenstein — Function

is_q_gorenstein( v::NormalToricVariety )

Checks if the normal toric variety v is Q-Gorenstein.

Examples

julia> is_q_gorenstein( projective_space( 2 ) )
true

JToric.is_fano — Function

is_fano( v::NormalToricVariety )

Checks if the normal toric variety v is fano.

Examples

julia> is_fano( projective_space( 2 ) )
true

Attributes of toric varieties

JToric.affine_open_covering — Function

affine_open_covering( v::NormalToricVariety )

Computes an affine open cover of the normal toric variety v, i.e. returns a list of affine toric varieties.

JToric.cox_ring — Function

cox_ring( v::NormalToricVariety )

Computes the Cox ring of the normal toric variety v.

cox_ring( v::NormalToricVariety, name::String )

Compute the Cox ring of the normal toric variety v, using name as label for the homogeneous coordinates.

JToric.list_of_variables_of_cox_ring — Function

list_of_variables_of_cox_ring( v::NormalToricVariety )

Computes the list of homogeneous variables of the Cox ring of the normal toric variety v.

JToric.class_group — Function

class_group( v::NormalToricVariety )

Computes the class group of the normal toric variety v.

JToric.torus_invariant_divisor_group — Function

torus_invariant_divisor_group( v::NormalToricVariety )

Computes the torus invariant divisor class group of the normal toric variety v.

JToric.map_from_character_to_principal_divisor — Function

map_from_character_to_principal_divisor( v::NormalToricVariety )

Computes the map from the character lattice to the principal divisors of the normal toric variety v.

JToric.map_from_weil_divisors_to_class_group — Function

map_from_weil_divisors_to_class_group( v::NormalToricVariety )

Computes the map from the Weil divisors to the Class group of the normal toric variety v.

JToric.dim — Function

dim( v::NormalToricVariety )

Computes the dimension of the normal toric variety v.

JToric.dim_of_torusfactor — Function

dim_of_torusfactor( v::NormalToricVariety )

Computes the dimension of the torus factor of the normal toric variety v.

JToric.coordinate_ring_of_torus — Function

coordinate_ring_of_torus( v::NormalToricVariety )

Computes the coordinate ring of the torus of the normal toric variety v.

coordinate_ring_of_torus( v::NormalToricVariety, names::Vector{String} )

Compute the coordinate ring of the torus factor of the normal toric variety v, using names as label for the homogeneous coordinates.

JToric.list_of_variables_of_coordinate_ring_of_torus — Function

list_of_variables_of_coordinate_ring_of_torus( v::NormalToricVariety )

Computes the list of homogeneous coordinates of the coordinate ring of the torus of the normal toric variety v.

JToric.is_product_of — Function

is_product_of( v::NormalToricVariety )

Identifies the factors from which the normal toric variety v has been constructed in GAP.

JToric.factors — Function

factors( v::NormalToricVariety )

Identifies the factors from which the normal toric variety v has been constructed in GAP.

JToric.character_lattice — Function

character_lattice( v::NormalToricVariety )

Computes the character lattice of the normal toric variety v.

JToric.torus_invariant_prime_divisors — Function

torus_invariant_prime_divisors( v::NormalToricVariety )

Computes the torus invariant prime divisors of the normal toric variety v.

JToric.irrelevant_ideal — Function

irrelevant_ideal( v::NormalToricVariety )

Computes the irrelevant ideal of the normal toric variety v.

JToric.stanley_reisner_ideal — Function

stanley_reisner_ideal( v::NormalToricVariety )

Computes the Stanley-Reisner ideal of the normal toric variety v.

JToric.morphism_from_cox_variety — Function

morphism_from_cox_variety( v::NormalToricVariety )

Computes the morphism from the Cox variety of the normal toric variety v.

JToric.cox_variety — Function

cox_variety( v::NormalToricVariety )

Computes the Cox variety of the normal toric variety v.

JToric.fan_of_variety — Function

fan_of_variety( v::NormalToricVariety )

Computes the fan of the normal toric variety v.

JToric.fan — Function

fan( v::NormalToricVariety )

A convenience method for fan_of_variety( v::NormalToricVariety ).

JToric.cartier_torus_invariant_divisor_group — Function

cartier_torus_invariant_divisor_group( v::NormalToricVariety )

Computes the group of Cartier and torus invariant divisors of the normal toric variety v.

JToric.picard_group — Function

picard_group( v::NormalToricVariety )

Computes the Picard group of the normal toric variety v.

JToric.name_of_variety — Function

name_of_variety( v::NormalToricVariety )

Returns the name of the normal toric variety v, if set. Otherwise returns "No name set for this variety".

JToric.zariski_cotangent_sheaf — Function

zariski_cotangent_sheaf( v::NormalToricVariety )

Returns the Zariski cotangent sheaf of the normal toric variety v.

JToric.cotangent_sheaf — Function

cotangent_sheaf( v::NormalToricVariety )

Returns the cotangent sheaf of the normal toric variety v.

JToric.euler_characteristic — Function

euler_characteristic( v::NormalToricVariety )

Computes the Euler characteristic of the normal toric variety v.

JToric.weil_divisors_of_variety — Function

weil_divisors_of_variety( v::NormalToricVariety )

Compute the Weil divisors of the normal toric variety v.

JToric.zariski_cotangent_sheaf_via_euler_sequence — Function

zariski_cotangent_sheaf_via_euler_sequence( v::NormalToricVariety )

Computes the Zariski cotangent sheaf of the normal toric variety v via the Euler sequence.

JToric.zariski_cotangent_sheaf_via_poincare_residue_map — Function

zariski_cotangent_sheaf_via_poincare_residue_map( v::NormalToricVariety )

Computes the Zariski cotangent sheaf of the normal toric variety v via the Poincare residue map.

JToric.nef_cone — Function

nef_cone( v::NormalToricVariety )

Computes the nef cone of the normal toric variety v.

JToric.mori_cone — Function

mori_cone( v::NormalToricVariety )

Computes the mori cone of the normal toric variety v.

Methods of toric varieties

JToric.set_name_of_variety — Function

set_name_of_variety( v::NormalToricVariety, name::String )

Sets the name of the normal toric variety v to name.

JToric.coordinate_ring_of_torus — Method

coordinate_ring_of_torus( v::NormalToricVariety, names::Vector{String} )

Compute the coordinate ring of the torus factor of the normal toric variety v, using names as label for the homogeneous coordinates.

JToric.cox_ring — Method

cox_ring( v::NormalToricVariety, name::String )

Compute the Cox ring of the normal toric variety v, using name as label for the homogeneous coordinates.

Base.:* — Method

Base.:*( v::NormalToricVariety, w::NormalToricVariety )

Compute the direct product of the normal toric varieties v and w.

Examples

julia> projective_space( 2 ) * projective_space( 2 )
NormalToricVariety(GAP: <A projective toric variety of dimension 4 which is a product of 2 toric varieties>, Polymake.BigObjectAllocated(Ptr{Nothing} @0x00005645a1a00930))

JToric.character_to_rational_function — Method

character_to_rational_function( l::Vector{Int}, v::NormalToricVariety )

Turn the character l of the normal toric variety v into a rational function.

JToric.blowup_on_ith_minimal_torus_orbit — Method

blowup_on_ith_minimal_torus_orbit( v::NormalToricVariety, i::Int )

Compute the blow up of the normal toric variety v on the i-th minimal torus orbit.

JToric.ith_betti_number — Method

ith_betti_number( v::NormalToricVariety, i::Int )

Compute the i-th Betti number of the normal toric variety v.

JToric.nr_of_q_rational_points — Method

nr_of_q_rational_points( v::NormalToricVariety, i::Int )

Compute the number of q-rational points of the normal toric variety v.

Toric Divisors

Constructors

JToric.create_divisor — Function

create_divisor( c::Vector{Int}, v::NormalToricVariety )

Construct the torus invariant divisor on the normal toric variety v as linear combination of the torus invariant prime divisors of v. The coefficients of thi linear combination are passed as list of integers as first argument.

Examples

julia> create_divisor( [1,1,2], projective_space( 2 ) )
toric_divisor(GAP: <A divisor of a toric variety with coordinates ( 1, 1, 2 )>)

JToric.divisor_of_character — Function

divisor_of_character( c::Vector{Int}, v::NormalToricVariety )

Construct the torus invariant divisor on the normal toric variety v corresponding to the character c.

Examples

julia> divisor_of_character( [1,2], projective_space( 2 ) )
toric_divisor(GAP: <A principal divisor of a toric variety with coordinates ( -3, 2, 1 )>)

JToric.divisor_of_class — Function

divisor_of_class( v::NormalToricVariety, c::Vector{Int} )

Construct a torus invariant divisor on the normal toric variety v corresponding to the divisor class c.

Examples

julia> divisor_of_class( [1], projective_space( 2 ) )
toric_divisor(GAP: <A divisor of a toric variety with coordinates ( 1, 0, 0 )>)

Properties of toric divisors

JToric.is_cartier — Function

is_cartier( d::ToricDivisor )

Checks if the divisor d is Cartier.

JToric.is_principal — Function

is_principal( d::ToricDivisor )

Checks if the divisor d is principal.

JToric.is_primedivisor — Function

is_primedivisor( d::ToricDivisor )

Checks if the divisor d is prime.

JToric.is_basepoint_free — Function

is_basepoint_free( d::ToricDivisor )

Checks if the divisor d is basepoint free.

JToric.is_ample — Function

is_ample( d::ToricDivisor )

Checks if the divisor d is ample.

JToric.is_very_ample — Function

is_very_ample( d::ToricDivisor )

For ample divisors d, this method checks if d is very ample.

JToric.is_nef — Function

is_nef( d::ToricDivisor )

Checks if the divisor d is numerically effective.