JToric – toric geometry in Julia

Introduction

Generalities

JToric – toric geometry in Julia

Introduction

The goal of the package JToric is to make computations on toric geometry accessible in the OSCAR computer algebra system. As such, we follow the book Toric Varieties by David A. Cox, John B. Little and Henry K. Schenck]. Currently, this project is work-in-progress. Our goal is to ensure that one can perform all computations of Appendix B.

Generalities

JToric.version — Constant

version

The version number of the loaded JToric. Please mention this number in any bug report.