JToric – toric geometry in Julia

JToric – toric geometry in Julia

Goal

Toric Varieties

Toric Divisors

Goal

The goal of the package JToric is to make computations on toric geometry accessible in the OSCAR computer algebra system. As such, we follow the book Toric Varieties by David A. Cox, John B. Little and Henry K. Schenck]. Currently, this project is work-in-progress. Our goal is to ensure that one can perform all computations of Appendix B.

Generalities

JToric.version — Constant

version

The version number of the loaded JToric. Please mention this number in any bug report.

Toric Varieties

Constructors

JToric.AffineNormalToricVariety — Method

AffineNormalToricVariety(C::Cone)

Construct the affine normal toric variety $U_{C}$ corresponding to a polyhedral cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = Oscar.positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

JToric.NormalToricVariety — Method

NormalToricVariety(C::Cone)

Construct the (affine) normal toric variety $X_{\Sigma}$ corresponding to a polyhedral fan $\Sigma = C$ consisting only of the cone C.

Examples

Set C to be the positive orthant in two dimensions.

julia> C = Oscar.positive_hull([1 0; 0 1])
A polyhedral cone in ambient dimension 2
julia> ntv = NormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

JToric.NormalToricVariety — Method

NormalToricVariety(PF::PolyhedralFan)

Construct the normal toric variety $X_{PF}$ corresponding to a polyhedral fan PF.

Examples

Take PF to be the normal fan of the square.

julia> square = Oscar.cube(2)
A polyhedron in ambient dimension 2

julia> nf = Oscar.normal_fan(square)
A polyhedral fan in ambient dimension 2

julia> ntv = NormalToricVariety(nf)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

JToric.NormalToricVariety — Method

NormalToricVariety( r::Matrix{Int}, c::Vector{Vector{Int}} )

Construct the normal toric variety whose fan has ray generators r and maximal cones c.

Examples

julia> NormalToricVariety( [-1 5; 0 1; 1 0; 0 -1], [[1,2],[2,3],[3,4],[4,1]] )
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Oscar.Geometry.projective_space — Function

projective_space( d::Int )

Construct the projective space of dimension d.

Examples

julia> projective_space( 2 )
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

JToric.hirzebruch_surface — Function

hirzebruch_surface( r::Int )

Constructs the r-th Hirzebruch surface.

Examples

julia> hirzebruch_surface( 5 )
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

JToric.del_pezzo — Function

delPezzo( b::Int )

Constructs the delPezzo surface with b blowups for b at most 3.

Examples

julia> del_pezzo( 3 )
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

Conversion among GAP and Polymake toric varieties

Missing docstring.

Missing docstring for ntv_gap2polymake. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ntv_polymake2gap. Check Documenter's build log for details.

Properties of toric varieties

Hecke.isnormal — Function

 isnormal(K::AnticNumberField) -> Bool

Returns true if $K$ is a normal extension of $\mathbb Q$, false otherwise.

isnormal(C::ClassField) -> Bool

For a class field $C$ defined over a normal base field $k$, decide if $C$ is normal over $Q$.

isnormal(a::AlgAssAbsOrdIdl) -> Bool

Returns true if $a$ is a normal ideal and false otherwise.

isnormal(G::T, H::T) where T <: GAPGroup

Return whether the subgroup H is normal in G, i. e., H is invariant under conjugation with elements of G.

isnormal(A::MPolyQuo)

Given an affine algebra A over a perfect field, return true if A is normal, false otherwise.

CAVEAT: The function computes the normalization of A. This may take some time.

isnormal(P::Polyhedron)

Check whether P is normal.

Examples

The 3-cube is normal.

julia> C = cube(3)
A polyhedron in ambient dimension 3

julia> isnormal(C)
true

But this pyramid is not:

julia> P = convex_hull([0 0 0; 0 1 1; 1 1 0; 1 0 1]);

julia> isnormal(P)
false

isnormal( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is normal. (This function is somewhat tautological at this point.)

Examples

julia> isnormal(projective_space( 2 ))
true

JToric.isaffine — Function

isaffine( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is affine.

Examples

julia> isaffine( projective_space( 2 ) )
false

JToric.isprojective — Function

isprojective( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is projective, i.e. if the fan of v is the the normal fan of a polytope.

Examples

julia> isprojective( projective_space( 2 ) )
true

Hecke.issmooth — Function

issmooth(C::AffinePlaneCurve{S}, P::Point{S}) where S <: FieldElem

Throw an error if P is not a point of C, return false if P is a singular point of C, and true if P is a smooth point of C.

Example

julia> R, (x, y) = PolynomialRing(QQ, ["x", "y"])
(Multivariate Polynomial Ring in x, y over Rational Field, fmpq_mpoly[x, y])

julia> C = Oscar.AffinePlaneCurve(x^2*(x+y)*(y^3-x^2))
Affine plane curve defined by -x^5 - x^4*y + x^3*y^3 + x^2*y^4

julia> P = Oscar.Point([QQ(0), QQ(0)])
Point with coordinates fmpq[0, 0]

julia> Oscar.issmooth(C, P)
false

issmooth(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem

Throw an error if P is not a point of C, return false if P is a singular point of C, and true if P is a smooth point of C.

Example

julia> S, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> T, _ = grade(S)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])

julia> C = Oscar.ProjPlaneCurve(x^2*(x+y)*(y^3-x^2*z))
Projective plane curve defined by -x^5*z - x^4*y*z + x^3*y^3 + x^2*y^4

julia> PP = projective_space(QQ, 2)
(Projective space of dim 2 over Rational Field
, MPolyElem_dec{fmpq, fmpq_mpoly}[x[0], x[1], x[2]])

julia> P = Oscar.Geometry.ProjSpcElem(PP[1], [QQ(0), QQ(0), QQ(1)])
(0 : 0 : 1)

julia> Oscar.issmooth(C, P)
false

issmooth(P::Polyhedron)

Check whether P is smooth.

Examples

A cube is always smooth.

julia> C = cube(8);

julia> issmooth(C)
true

issmooth(PF::PolyhedralFan)

Determine whether PF is smooth.

Examples

Even though the cones of this fan cover the positive orthant together, one of these und thus the whole fan is not smooth.

julia> PF = PolyhedralFan([0 1; 2 1; 1 0], IncidenceMatrix([[1, 2], [2, 3]]));

julia> issmooth(PF)
false

issmooth( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is smooth.

Examples

julia> issmooth( projective_space( 2 ) )
true

Oscar.iscomplete — Function

iscomplete(PF::PolyhedralFan)

Determine whether PF is complete, i.e. its support, the set-theoretic union of its cones, covers the whole space.

Examples

Normal fans of polytopes are complete.

julia> iscomplete(normal_fan(cube(3)))
true

iscomplete( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is complete.

Examples

julia> iscomplete( projective_space( 2 ) )
true

JToric.has_torusfactor — Function

has_torusfactor( v::AbstractNormalToricVariety )

Checks if the normal toric variety v has a torus factor.

Examples

julia> has_torusfactor( projective_space( 2 ) )
false

JToric.is_orbifold — Function

is_orbifold( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is an orbifold.

Examples

julia> is_orbifold( projective_space( 2 ) )
true

JToric.issimplicial — Function

issimplicial( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is simplicial. Hence, this function works just as is_orbifold. It is implemented for user convenience.

Examples

julia> issimplicial( projective_space( 2 ) )
true

Missing docstring.

Missing docstring for is_isomorphic_to_projective_space. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_direct_product_of_projective_spaces. Check Documenter's build log for details.

JToric.is_gorenstein — Function

is_gorenstein( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is Gorenstein.

Examples

julia> is_gorenstein( projective_space( 2 ) )
true

JToric.is_q_gorenstein — Function

is_q_gorenstein( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is Q-Gorenstein.

Examples

julia> is_q_gorenstein( projective_space( 2 ) )
true

JToric.is_fano — Function

is_fano( v::AbstractNormalToricVariety )

Checks if the normal toric variety v is fano.

Examples

julia> is_fano( projective_space( 2 ) )
true

Attributes of toric varieties

Missing docstring.

Missing docstring for affine_open_covering. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cox_ring. Check Documenter's build log for details.

Missing docstring.

Missing docstring for list_of_variables_of_cox_ring. Check Documenter's build log for details.

Hecke.class_group — Function

class_group(K::AnticNumberField) -> GrpAbFinGen, Map

Shortcut for class_group(maximal_order(K)): returns the class group as an abelian group and a map from this group to the set of ideals of the maximal order.

class_group(O::NfOrd; bound = -1, method = 3, redo = false, large = 1000) -> GrpAbFinGen, Map

Returns a group $A$ and a map $f$ from $A$ to the set of ideals of $O$. The inverse of the map is the projection onto the group of ideals modulo the group of principal ideals. redo allows to trigger a re-computation, thus avoiding the cache. bound, when given, is the bound for the factor base.

Missing docstring.

Missing docstring for torus_invariant_divisor_group. Check Documenter's build log for details.

Missing docstring.

Missing docstring for map_from_character_to_principal_divisor. Check Documenter's build log for details.

Missing docstring.

Missing docstring for map_from_weil_divisors_to_class_group. Check Documenter's build log for details.

AbstractAlgebra.Generic.dim — Function

dim(Y::YoungTableau) -> BigInt

Return the dimension (using hook-length formula) of the irreducible representation of permutation group $S_n$ associated the partition Y.part.

Since the computation overflows easily BigInt is returned. You may perform the computation of the dimension in different type by calling dim(Int, Y).

Examples

julia> dim(YoungTableau([4,3,1]))
70

julia> dim(YoungTableau([3,1])) # the regular representation of S_4
3

dim(M::FreeModule{T}) where T <: FieldElement

Return the dimension of the given vector space.

dim(N::Submodule{T}) where T <: FieldElement

Return the dimension of the given vector subspace.

dim(N::QuotientModule{T}) where T <: FieldElement

Return the dimension of the given vector quotient space.

dim(V::AbsSpace) -> Int

Return the dimension of the space V.

dim(S::ZpGenus) -> fmpz

Return the dimension of this genus.

dim(G::ZGenus) -> Int

Return the dimension of this genus.

dim(I::MPolyIdeal)

Return the Krull dimension of I.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> I = ideal(R, [y-x^2, x-z^3])
ideal(-x^2 + y, x - z^3)

julia> dim(I)
1

dim(a::MPolyQuoIdeal)

Return the Krull dimension of a.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"], ordering=:degrevlex)
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, q = quo(R, ideal(R, [x^3*y^2-y^3*x^2, x*y^4-x*y^2]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(x^3*y^2 - x^2*y^3, x*y^4 - x*y^2) defined by a julia-function with inverse
)

julia> a = ideal(Q,[x^3*y^4-x+y, x*y+y^2*x])
ideal(x^3*y^4 - x + y, x*y^2 + x*y)

julia> dim(a)
1

dim(Q::MPolyQuo)

Return the Krull dimension of the quotient ring Q.

Examples

julia> R, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> Q, _ = quo(R, ideal(R, [y-x^2, x-z^3]))
(Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3), Map from
Multivariate Polynomial Ring in x, y, z over Rational Field to Quotient of Multivariate Polynomial Ring in x, y, z over Rational Field by ideal(-x^2 + y, x - z^3) defined by a julia-function with inverse
)

julia> dim(Q)
1

dim(C::Cone)

Return the dimension of C.

Examples

The cone C in this example is 2-dimensional within a 3-dimensional ambient space.

julia> C = Cone([1 0 0; 1 1 0; 0 1 0]);

julia> dim(C)
2

dim(P::Polyhedron)

Return the dimension of P.

Examples

julia> V = [1 2 3; 1 3 2; 2 1 3; 2 3 1; 3 1 2; 3 2 1];

julia> P = convex_hull(V);

julia> dim(P)
2

dim(PF::PolyhedralFan)

Return the dimension of PF.

Examples

This fan in the plane contains a 2-dimensional cone and is thus 2-dimensional itself.

julia> PF = PolyhedralFan([1 0; 0 1; -1 -1], IncidenceMatrix([[1, 2], [3]]));

julia> dim(PF)
2

dim( v::AbstractNormalToricVariety )

Computes the dimension of the normal toric variety v.

JToric.dim_of_torusfactor — Function

dim_of_torusfactor( v::AbstractNormalToricVariety )

Computes the dimension of the torus factor of the normal toric variety v.

Missing docstring.

Missing docstring for coordinate_ring_of_torus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for list_of_variables_of_coordinate_ring_of_torus. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_product_of. Check Documenter's build log for details.

Missing docstring.

Missing docstring for factors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for character_lattice. Check Documenter's build log for details.

Missing docstring.

Missing docstring for torus_invariant_prime_divisors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for irrelevant_ideal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for stanley_reisner_ideal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for morphism_from_cox_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cox_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fan_of_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for fan. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cartier_torus_invariant_divisor_group. Check Documenter's build log for details.

Hecke.picard_group — Function

picard_group(O::NfOrd) -> GrpAbFinGen, MapClassGrp

Returns the Picard group of $O$ and a map from the group in the set of (invertible) ideals of $O$.

picard_group(O::AlgAssAbsOrd, prepare_ref_disc_log::Bool = false)
  -> GrpAbFinGen, MapPicardGroup

Given an order $O$ in a commutative algebra over $\mathbb Q$, this function returns the picard group of $O$. If prepare_ref_disc_log is true, then (possibly expensive) preparations for the computation of refined discrete logarithms in non maximal orders are done.

Missing docstring.

Missing docstring for name_of_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for zariski_cotangent_sheaf. Check Documenter's build log for details.

Missing docstring.

Missing docstring for cotangent_sheaf. Check Documenter's build log for details.

JToric.euler_characteristic — Function

euler_characteristic( v::AbstractNormalToricVariety )

Computes the Euler characteristic of the normal toric variety v.

Missing docstring.

Missing docstring for weil_divisors_of_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for zariski_cotangent_sheaf_via_euler_sequence. Check Documenter's build log for details.

Missing docstring.

Missing docstring for zariski_cotangent_sheaf_via_poincare_residue_map. Check Documenter's build log for details.

JToric.nef_cone — Function

nef_cone( v::NormalToricVariety )

Computes the nef cone of the normal toric variety v.

JToric.mori_cone — Function

mori_cone( v::NormalToricVariety )

Computes the mori cone of the normal toric variety v.

JToric.toric_ideal_binomial_generators — Method

toric_ideal_binomial_generators(antv::AffineNormalToricVariety)

Get the exponent vectors corresponding to the generators of the toric ideal associated to the affine normal toric variety antv.

Examples

Take the cyclic quotient singularity corresponding to the pair of integers (2,5).

julia> C = Oscar.positive_hull([-2 5; 1 0])
A polyhedral cone in ambient dimension 2

julia> antv = AffineNormalToricVariety(C)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> toric_ideal_binomial_generators(antv)
pm::Matrix<long>
-1 -1 2 1
-1 0 3 -1
0 -1 -1 2

Methods of toric varieties

Missing docstring.

Missing docstring for set_name_of_variety. Check Documenter's build log for details.

Missing docstring.

Missing docstring for coordinate_ring_of_torus( v::NormalToricVariety, names::Vector{String} ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for cox_ring( v::NormalToricVariety, name::String ). Check Documenter's build log for details.

Base.:* — Method

*(x, y...)

Multiplication operator. x*y*z*... calls this function with all arguments, i.e. *(x, y, z, ...).

Examples

julia> 2 * 7 * 8
112

julia> *(2, 7, 8)
112

Missing docstring.

Missing docstring for character_to_rational_function( l::Vector{Int}, v::NormalToricVariety ). Check Documenter's build log for details.

Missing docstring.

Missing docstring for blowup_on_ith_minimal_torus_orbit( v::NormalToricVariety, i::Int ). Check Documenter's build log for details.

JToric.ith_betti_number — Method

ith_betti_number( v::AbstractNormalToricVariety, i::Int )

Compute the i-th Betti number of the normal toric variety v.

Missing docstring.

Missing docstring for nr_of_q_rational_points( v::NormalToricVariety, i::Int ). Check Documenter's build log for details.

Toric Divisors

Constructors

JToric.ToricDivisor — Type

ToricDivisor(coeffs::AbstractVector, v::AbstractNormalToricVariety )

Construct the torus invariant divisor on the normal toric variety v as linear combination of the torus invariant prime divisors of v. The coefficients of thi linear combination are passed as list of integers as first argument.

Examples

julia> show( ToricDivisor( [1,1,2], projective_space( 2 ) ) )
A torus invariant divisor on a normal toric variety

Missing docstring.

Missing docstring for divisor_of_character. Check Documenter's build log for details.

Missing docstring.

Missing docstring for divisor_of_class. Check Documenter's build log for details.

Properties of toric divisors

JToric.isample — Function

isample(td::ToricDivisor)

Determine whether the toric divisor td is ample.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isample(td)
false

JToric.isbasepoint_free — Function

isbasepoint_free(td::ToricDivisor)

Determine whether the toric divisor td is basepoint free.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isbasepoint_free(td)
true

JToric.iscartier — Function

is_cartier( d::ToricDivisor )

Checks if the divisor d is Cartier.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> iscartier(td)
true

JToric.iseffective — Method

iseffective(td::ToricDivisor)

Determine whether the toric divisor td is effective.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> iseffective(td)
true

Hecke.isintegral — Method

isintegral(td::ToricDivisor)

Determine whether the toric divisor td is integral.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isintegral(td)
true

JToric.isnef — Function

isnef(td::ToricDivisor)

Determine whether the toric divisor td is nef.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isnef(td)
true

Missing docstring.

Missing docstring for is_primedivisor. Check Documenter's build log for details.

Hecke.isprincipal — Function

isprincipal(A::NfOrdIdl) -> Bool, NfOrdElem
isprincipal(A::NfOrdFracIdl) -> Bool, NfOrdElem

Tests if $A$ is principal and returns $(\mathtt{true}, \alpha)$ if $A = \langle \alpha\rangle$ or $(\mathtt{false}, 1)$ otherwise.

isprincipal(D::ProjCurveDivisor{S}) where S <: FieldElem

Return true if the divisor D is principal, and false otherwise

Example

julia> S, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])

julia> T, _ = grade(S)
(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
  x -> [1]
  y -> [1]
  z -> [1], MPolyElem_dec{fmpq, fmpq_mpoly}[x, y, z])

julia> C = Oscar.ProjPlaneCurve(T(y^2*z - x*(x-z)*(x+3*z)))
Projective plane curve defined by -x^3 - 2*x^2*z + 3*x*z^2 + y^2*z

julia> PP = projective_space(QQ, 2)
(Projective space of dim 2 over Rational Field
, MPolyElem_dec{fmpq, fmpq_mpoly}[x[0], x[1], x[2]])

julia> P = Oscar.Geometry.ProjSpcElem(PP[1], [QQ(0), QQ(1), QQ(0)])
(0 : 1 : 0)

julia> E = Oscar.ProjCurveDivisor(C, P)
(0 : 1 : 0)

julia> Oscar.isprincipal(E)
false

isprincipal(td::ToricDivisor)

Determine whether the toric divisor td is principal.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isprincipal(td)
false

JToric.isq_cartier — Function

isq_cartier(td::ToricDivisor)

Determine whether the toric divisor td is Q-Cartier.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isq_cartier(td)
true

JToric.isvery_ample — Function

isvery_ample(td::ToricDivisor)

Determine whether the toric divisor td is very ample.

Examples

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isvery_ample(td)
false

JToric.polyhedron_of_divisor — Method

polyhedron_of_divisor(td::ToricDivisor)

Construct the polyhedron $P_D$ of a torus invariant divisor $D:=td$ as in 4.3.2 of CLS11. The lattice points of this polyhedron correspond to the global sections of the divisor.

Examples

The polyhedron of the divisor with all coefficients equal to zero is a point, if the ambient variety is complete. Changing the coefficients corresponds to moving hyperplanes. One direction moves the hyperplane away from the origin, the other moves it across. In the latter case there are no global sections anymore and the polyhedron becomes empty.

julia> H = hirzebruch_surface(4)
A normal toric variety corresponding to a polyhedral fan in ambient dimension 2

julia> td0 = ToricDivisor([0,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isfeasible(polyhedron_of_divisor(td0))
true

julia> dim(polyhedron_of_divisor(td0))
0

julia> td1 = ToricDivisor([1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isfeasible(polyhedron_of_divisor(td1))
true

julia> td2 = ToricDivisor([-1,0,0,0], H)
A torus invariant divisor on a normal toric variety

julia> isfeasible(polyhedron_of_divisor(td2))
false